% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{my_data}
\alias{my_data}
\title{Example Data: my_data}
\format{
A data frame with 10 rows and 2 columns:
\describe{
  \item{x}{A numeric vector of values 1 to 10.}
  \item{y}{A numeric vector of random values.}
}
}
\source{
Created with data.frame function.
}
\usage{
data(my_data)
}
\description{
Example Data: my_data
}
\examples{
# Load the data
data(my_data)
# View the first few rows
head(my_data)
}
\keyword{datasets}
