% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{plot_line}
\alias{plot_line}
\title{Plot a line}
\usage{
plot_line(m, b, x = c(1:100))
}
\arguments{
\item{m}{slope of the line}

\item{b}{intercept of the line}

\item{x}{vector of x values}
}
\value{
A plot of the line
}
\description{
Plot a line with slope m and intercept b
}
\examples{
plot_line(2,3)
plot_line(0,2,x=c(1:10))

}
