% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{plot_square}
\alias{plot_square}
\title{Plot a square
Plot a square with a and c}
\usage{
plot_square(a, c, x = c(1:100))
}
\arguments{
\item{a}{coefficient of x}

\item{c}{constant term}

\item{x}{vector of x values}
}
\value{
A plot of the square
}
\description{
Plot a square
Plot a square with a and c
}
\examples{
plot_square(2,3)
plot_square(0,2,x=c(1:10))

}
